attach homepage and booking
