+++ src/routes/routesConfig.tsx
@@ -14,7 +14,6 @@ const LoginPage = lazy(() => import('views/auth/Login/LoginPage'));
 const RegisterPage = lazy(() => import('views/auth/Register/RegisterPage'));
 const ChangePasswordPage = lazy(() => import('views/auth/ChangePassword/ChangePasswordPage'));
 const ForgotPasswordPage = lazy(() => import('views/auth/ForgotPassword/ForgotPasswordPage'));
-const ListTripRoute = lazy(() => import('views/Booking/listTripRoute'));
 
 /* eslint-disable  @typescript-eslint/naming-convention  */
 export enum ROUTER_PATH {
@@ -54,15 +53,11 @@ export const routesConfig: RouteObject[] = [
         children: [
           {
             path: '/',
-            element: <HomePage/>,
+            element: <HomePage />,
           },
           {
             path: '/booking',
             element: <BookingPage />,
-          },
-          {
-            path: '/booking/trip-routes',
-            element: <ListTripRoute />,
           },
         ],
       },
